feat : hotfix datatype
because of overflow, fix type int to long.

+++ src/main/java/lotto/Rank.java
@@ -17,7 +17,7 @@ public enum Rank {
         this.displayName = displayName;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
